[Chore] Relocate InvitedDashboards folder
